refactor: move context file
