+++ src/components/ContextProvider/Context.tsx
@@ -6,8 +6,8 @@ import {
   createContext,
   useState,
 } from 'react';
-import { IGif } from './types';
-import { BASE_LIMIT } from './api';
+import { IGif } from '../../types';
+import { BASE_LIMIT } from '../../api';
 
 interface IContextProvider {
   children: ReactNode;
